requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql          
python-dotenv    # to load environment variables
alembic          # optional, for migrations
